separated Generator into its own module
